+++ packages/calendar/src/components/input-types/date-time-input/DateTimeInput.tsx
@@ -29,6 +29,7 @@ export interface DateTimeInputProps extends ValueAndOnValueChangeProps<Date> {
   onEsc?: () => void;
   onEnter?: () => void;
   onBlur?: () => void;
+  autoFocus?: boolean;
 }
 
 export const DateTimeInput: React.FC<DateTimeInputProps> = ({
@@ -37,6 +38,7 @@ export const DateTimeInput: React.FC<DateTimeInputProps> = ({
   onEnter,
   onEsc,
   onBlur,
+  autoFocus,
 }) => {
   const popoverRef = useRef(null);
   const containerRef = useRef(null);
@@ -145,6 +147,7 @@ export const DateTimeInput: React.FC<DateTimeInputProps> = ({
         }
       >
         <DualTextInput
+          autoFocusLeft={autoFocus}
           onEsc={onEsc}
           onEnter={onEnter}
           onBlur={onBlur}